+++ src/services/localforage.js
@@ -36,14 +36,23 @@ const localForagePlugin = (({ store }) => {
         storeName: store.$id
     });
 
+    // These are the state variables currently in the Pinia store
+    // We will remove the ones that are not in the store from indexedDB
+    const storeKeys = Object.keys(store.$state);
+
     // Load the persisted state from storage.
     // Called at application start, when the plugin is registered for the store.
     let persistedState = {};
     storage.iterate((value, key, iterationNumber) => {
         persistedState[key] = JSON.parse(value);
         // Remove keys with empty values, so default values can be used instead
-        if ( persistedState[key] == null || persistedState[key].length == 0 ) {
+        if (persistedState[key] == null || persistedState[key].length == 0) {
             delete persistedState[key];
+        }
+
+        // Keep indexedDB clean from old and now unused persisted state keys.
+        if (!storeKeys.includes(key)) {
+            storage.removeItem(key);
         }
     }).then(() => {
         // Restore the persisted state
@@ -60,7 +69,7 @@ const localForagePlugin = (({ store }) => {
         for (const element in state) {
             storage.setItem(element, JSON.stringify(state[element]));
         }
-    })
+    });
 })
 
 export {